# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Kaspi Pay POS Automation API",
    "version": "1.0.0",
    "description": "REST API for integrating POS systems with Kaspi Pay API (SMS Authorization, Invoicing, QR Payments, Operation History, and Refunds)."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Theater API Key used for authentication and permission gating"
      },
      "AdminBearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Master admin secret token check (ADMIN_SECRET_KEY)"
      },
      "TokenSN": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Token-SN",
        "description": "Session token SN obtained after OTP verification"
      },
      "VtokenSecret": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Vtoken-Secret",
        "description": "Encrypted session token secret"
      },
      "ProfileId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Profile-Id",
        "description": "Organization profile ID (optional, sent after auth)"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "example": "Description of the error"
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/history/operations": {
      "post": {
        "tags": ["History"],
        "summary": "Get Statements History",
        "description": "Retrieves the history of operations/statements for the specified period.",
        "security": [
          {
            "TokenSN": [],
            "VtokenSecret": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["endDate"],
                "properties": {
                  "endDate": {
                    "type": "string",
                    "description": "End date in YYYY-MM-DD",
                    "example": "2026-06-23"
                  },
                  "lastTransactionDate": {
                    "type": "string",
                    "description": "Date of last transaction for pagination",
                    "example": ""
                  },
                  "statementPeriodCode": {
                    "type": "integer",
                    "description": "Period code offset",
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operations list returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "Id": {
                            "type": "integer",
                            "example": 11223344
                          },
                          "Amount": {
                            "type": "number",
                            "example": 5000
                          },
                          "Type": {
                            "type": "string",
                            "example": "Sale"
                          },
                          "CreatedDate": {
                            "type": "string",
                            "example": "2026-06-23T11:45:00.000"
                          },
                          "ConsumerName": {
                            "type": "string",
                            "example": "АЛИХАН Б."
                          },
                          "Status": {
                            "type": "string",
                            "example": "Confirmed"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing endDate parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "endDate required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
    "/api/history/details": {
      "post": {
        "tags": ["History"],
        "summary": "Get Operation Detail",
        "description": "Retrieves precise transaction details (payment types, items, commissions).",
        "security": [
          {
            "TokenSN": [],
            "VtokenSecret": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["id"],
                "properties": {
                  "id": {
                    "type": "integer",
                    "description": "Operation ID",
                    "example": 123456
                  },
                  "operationMethod": {
                    "type": "integer",
                    "example": 0
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Detailed payload returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "StatusCode": {
                      "type": "integer",
                      "example": 0
                    },
                    "Data": {
                      "type": "object",
                      "properties": {
                        "Id": {
                          "type": "integer",
                          "example": 11223344
                        },
                        "Amount": {
                          "type": "number",
                          "example": 5000
                        },
                        "Type": {
                          "type": "string",
                          "example": "Sale"
                        },
                        "CreatedDate": {
                          "type": "string",
                          "example": "2026-06-23T11:45:00.000"
                        },
                        "ConsumerName": {
                          "type": "string",
                          "example": "АЛИХАН Б."
                        },
                        "ConsumerCardMask": {
                          "type": "string",
                          "example": "4405********1122"
                        },
                        "Commission": {
                          "type": "number",
                          "example": 47.5
                        },
                        "NetAmount": {
                          "type": "number",
                          "example": 4952.5
                        },
                        "ReceiptUrl": {
                          "type": "string",
                          "example": "https://pay.kaspi.kz/receipt/..."
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Missing id parameter",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "id required"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API Key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Invalid API Key"
                }
              }
            }
          },
          "404": {
            "description": "Operation details not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Operation details not found"
                }
              }
            }
          },
          "500": {
            "description": "Server or database error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": "Database connection timeout"
                }
              }
            }
          }
        }
      }
    },
...
  }
}